workers_statistics/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.135.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pandas-stubs==3.0.0.260204
pluggy==1.6.0
psycopg==3.3.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
uvicorn==0.42.0
